Split train and predict
